+++ src/org/pushingpixels/substance/internal/contrib/randelshofer/quaqua/svg/ic_grid_on_black_24px.java
@@ -11,7 +11,7 @@ import org.pushingpixels.lafwidget.icon.IsResizable;
 
 /**
  * This class has been automatically generated using <a
- * href="https://flamingo.dev.java.net">Flamingo SVG transcoder</a>.
+ * href="https://github.com/kirill-grouchnikov/flamingo">Flamingo SVG transcoder</a>.
  */
 public class ic_grid_on_black_24px implements Icon, UIResource, IsResizable, IsHiDpiAware {
 	/**
@@ -131,8 +131,8 @@ g.setTransform(defaultTransform_);
      * 
      * @return The X of the bounding box of the original SVG image.
      */
-    public static int getOrigX() {
-        return 2;
+    public static double getOrigX() {
+        return 2.0;
     }
 
     /**
@@ -140,8 +140,8 @@ g.setTransform(defaultTransform_);
      * 
      * @return The Y of the bounding box of the original SVG image.
      */
-    public static int getOrigY() {
-        return 2;
+    public static double getOrigY() {
+        return 2.0;
     }
 
 	/**
@@ -149,8 +149,8 @@ g.setTransform(defaultTransform_);
 	 * 
 	 * @return The width of the bounding box of the original SVG image.
 	 */
-	public static int getOrigWidth() {
-		return 20;
+	public static double getOrigWidth() {
+		return 20.0;
 	}
 
 	/**
@@ -158,8 +158,8 @@ g.setTransform(defaultTransform_);
 	 * 
 	 * @return The height of the bounding box of the original SVG image.
 	 */
-	public static int getOrigHeight() {
-		return 20;
+	public static double getOrigHeight() {
+		return 20.0;
 	}
 
 	/**
@@ -176,8 +176,8 @@ g.setTransform(defaultTransform_);
 	 * Creates a new transcoded SVG image.
 	 */
 	public ic_grid_on_black_24px() {
-        this.width = getOrigWidth();
-        this.height = getOrigHeight();
+        this.width = (int) getOrigWidth();
+        this.height = (int) getOrigHeight();
 	}
 
 	/*
