+++ src/org/pushingpixels/substance/internal/contrib/randelshofer/quaqua/svg/ic_palette_black_24px.java
@@ -11,7 +11,7 @@ import org.pushingpixels.lafwidget.icon.IsResizable;
 
 /**
  * This class has been automatically generated using <a
- * href="https://flamingo.dev.java.net">Flamingo SVG transcoder</a>.
+ * href="https://github.com/kirill-grouchnikov/flamingo">Flamingo SVG transcoder</a>.
  */
 public class ic_palette_black_24px implements Icon, UIResource, IsResizable, IsHiDpiAware {
 	/**
@@ -102,8 +102,8 @@ g.setTransform(defaultTransform_);
      * 
      * @return The X of the bounding box of the original SVG image.
      */
-    public static int getOrigX() {
-        return 3;
+    public static double getOrigX() {
+        return 3.0;
     }
 
     /**
@@ -111,8 +111,8 @@ g.setTransform(defaultTransform_);
      * 
      * @return The Y of the bounding box of the original SVG image.
      */
-    public static int getOrigY() {
-        return 3;
+    public static double getOrigY() {
+        return 3.0;
     }
 
 	/**
@@ -120,8 +120,8 @@ g.setTransform(defaultTransform_);
 	 * 
 	 * @return The width of the bounding box of the original SVG image.
 	 */
-	public static int getOrigWidth() {
-		return 18;
+	public static double getOrigWidth() {
+		return 18.0;
 	}
 
 	/**
@@ -129,8 +129,8 @@ g.setTransform(defaultTransform_);
 	 * 
 	 * @return The height of the bounding box of the original SVG image.
 	 */
-	public static int getOrigHeight() {
-		return 18;
+	public static double getOrigHeight() {
+		return 18.0;
 	}
 
 	/**
@@ -147,8 +147,8 @@ g.setTransform(defaultTransform_);
 	 * Creates a new transcoded SVG image.
 	 */
 	public ic_palette_black_24px() {
-        this.width = getOrigWidth();
-        this.height = getOrigHeight();
+        this.width = (int) getOrigWidth();
+        this.height = (int) getOrigHeight();
 	}
 
 	/*
